# contracts/openapi/tool-defect-api-v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "刀具缺陷检测系统接口",
    "version": "1.0.0",
    "description": "采集、业务查询、复核、数据集、训练、模型发布和内部推理接口的 v1 唯一契约源。"
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "地址由环境配置提供，示例域名不可用于生产"
    }
  ],
...
  "paths": {
...
    "/api/v1/review-tasks": {
      "get": {
        "operationId": "listReviewTasks",
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "UserOidc": [
              "review:read"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "../json-schema/common-v1.schema.json#/$defs/ReviewStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "按权限范围过滤的复核任务池",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewTaskPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/review-tasks/{review_task_id}": {
      "get": {
        "operationId": "getReviewWorkspace",
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "UserOidc": [
              "review:read"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReviewTaskId"
          }
        ],
        "responses": {
          "200": {
            "description": "按权限范围过滤的复核任务与检测证据工作区",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewWorkspace"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/review-tasks/{review_task_id}/claim": {
      "post": {
        "operationId": "claimReviewTask",
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "UserOidc": [
              "review:claim"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReviewTaskId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ActionRequest"
        },
        "responses": {
          "200": {
            "description": "任务已认领并返回租约",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewTask"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "claimReviewTask"
      }
    },
    "/api/v1/review-tasks/{review_task_id}/release": {
      "post": {
        "operationId": "releaseReviewTask",
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "UserOidc": [
              "review:claim"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReviewTaskId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ActionRequest"
        },
        "responses": {
          "200": {
            "description": "租约已释放",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewTask"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "releaseReviewTask"
      }
    },
    "/api/v1/review-tasks/{review_task_id}/submissions": {
      "post": {
        "operationId": "submitReview",
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "UserOidc": [
              "review:submit"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReviewTaskId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "不可变复核记录已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSubmissionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "submitReview"
      }
    },
    "/api/v1/review-tasks/{review_task_id}/annotation-upload-ticket": {
      "post": {
        "operationId": "createAnnotationUploadTicket",
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "UserOidc": [
              "review:annotate"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReviewTaskId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationUploadTicketRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "短时受控上传授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTicketResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "createAnnotationUploadTicket"
      }
    },
    "/api/v1/review-tasks/{review_task_id}/annotations/{image_id}/complete": {
      "post": {
        "operationId": "completeReviewAnnotation",
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "UserOidc": [
              "review:annotate"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReviewTaskId"
          },
          {
            "$ref": "#/components/parameters/ImageId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "人工掩膜已确认，重复确认相同内容返回相同结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectCompleteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "completeReviewAnnotation"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "EdgeDeviceMtls": {
        "type": "mutualTLS"
      },
      "EdgeOAuth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://identity.example.invalid/oauth2/token",
            "scopes": {
              "capture:write": "写入采集事件",
              "capture:read": "读取采集状态",
              "device:heartbeat": "上报设备心跳"
            }
          }
        }
      },
      "UserOidc": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://identity.example.invalid/.well-known/openid-configuration"
      },
      "ServiceMtls": {
        "type": "mutualTLS"
      },
      "ServiceOAuth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://identity.example.invalid/oauth2/token",
            "scopes": {
              "inference:callback": "提交推理执行结果",
              "runtime:read": "读取推理运行时健康"
            }
          }
        }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 256
        }
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^\\\"?[0-9]+\\\"?$"
        }
      },
...
      "ImageId": {
        "name": "image_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
        }
      },
...
      "ReviewTaskId": {
        "name": "review_task_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
        }
      },
...
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 512
        }
      },
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        }
      }
    },
    "requestBodies": {
      "ActionRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ActionRequest"
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "请求格式错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "身份认证失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "Forbidden": {
        "description": "没有资源或操作权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在或调用者不可见",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "Conflict": {
        "description": "幂等摘要或乐观锁冲突",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "ValidationError": {
        "description": "语义或状态校验失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ObjectCompleteRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "size_bytes",
          "sha256",
          "upload_receipt"
        ],
        "properties": {
          "size_bytes": {
            "type": "integer",
            "minimum": 1
          },
          "sha256": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Sha256"
          },
          "upload_receipt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          }
        }
      },
      "ObjectCompleteResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "image_id",
          "state",
          "sha256"
        ],
        "properties": {
          "image_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "state": {
            "const": "AVAILABLE"
          },
          "sha256": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Sha256"
          }
        }
      },
...
      "DetectionDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "capture",
          "detection",
          "attempts",
          "disposition_history",
          "images",
          "versions"
        ],
        "properties": {
          "capture": {
            "$ref": "#/components/schemas/CaptureStatusResponse"
          },
          "detection": {
            "$ref": "#/components/schemas/DetectionSummary"
          },
          "attempts": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "disposition_history": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "../json-schema/common-v1.schema.json#/$defs/ImageReference"
            }
          },
          "versions": {
            "type": "object"
          }
        }
      },
...
      "ActionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "client_request_id"
        ],
        "properties": {
          "client_request_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "reason": {
            "type": "string",
            "maxLength": 512
          }
        }
      },
      "ReviewTask": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "review_task_id",
          "capture_id",
          "status",
          "priority",
          "record_version"
        ],
        "properties": {
          "review_task_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "capture_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "status": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/ReviewStatus"
          },
          "priority": {
            "type": "string",
            "enum": [
              "P0",
              "P1",
              "P2",
              "P3"
            ]
          },
          "lease_expires_at": {
            "oneOf": [
              {
                "$ref": "../json-schema/common-v1.schema.json#/$defs/UtcTimestamp"
              },
              {
                "type": "null"
              }
            ]
          },
          "record_version": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "ReviewWorkspace": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "task",
          "evidence"
        ],
        "properties": {
          "task": {
            "$ref": "#/components/schemas/ReviewTask"
          },
          "evidence": {
            "$ref": "#/components/schemas/DetectionDetail"
          }
        }
      },
      "ReviewTaskPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "next_cursor",
          "has_more"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewTask"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "has_more": {
            "type": "boolean"
          }
        }
      },
      "ReviewSubmissionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "decision",
          "reason_code",
          "comment",
          "defect_type_codes",
          "annotation_image_id",
          "client_submitted_at"
        ],
        "properties": {
          "decision": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/BusinessDisposition"
          },
          "reason_code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "comment": {
            "type": "string",
            "maxLength": 2000
          },
          "defect_type_codes": {
            "type": "array",
            "uniqueItems": true,
            "maxItems": 32,
            "items": {
              "type": "string",
              "maxLength": 64
            }
          },
          "annotation_image_id": {
            "oneOf": [
              {
                "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_submitted_at": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/UtcTimestamp"
          }
        }
      },
      "ReviewSubmissionResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "review_record_id",
          "task_status",
          "business_disposition",
          "record_version"
        ],
        "properties": {
          "review_record_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "task_status": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/ReviewStatus"
          },
          "business_disposition": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/BusinessDisposition"
          },
          "record_version": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "AnnotationUploadTicketRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "media_type",
          "size_bytes",
          "sha256",
          "width",
          "height"
        ],
        "properties": {
          "media_type": {
            "const": "image/png"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 1
          },
          "sha256": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Sha256"
          },
          "width": {
            "type": "integer",
            "minimum": 1,
            "maximum": 32768
          },
          "height": {
            "type": "integer",
            "minimum": 1,
            "maximum": 32768
          }
        }
      },
      "UploadTicketResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "image_id",
          "upload"
        ],
        "properties": {
          "image_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "upload": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "method",
              "url",
              "headers",
              "expires_at"
            ],
            "properties": {
              "method": {
                "const": "PUT"
              },
              "url": {
                "type": "string",
                "format": "uri",
                "pattern": "^https://"
              },
              "headers": {
                "type": "object",
                "maxProperties": 8
              },
              "expires_at": {
                "$ref": "../json-schema/common-v1.schema.json#/$defs/UtcTimestamp"
              }
            }
          }
        }
      },
...
    }
  },
...
}